pyproject.toml
[project]
name = "telegrammboticecreamshop"
version = "0.1.0"
description = ""
authors = [
    {name = "george2066",email = "george2066@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot (>=21.11.1,<22.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
asyncpg-stubs = "^0.30.0"

[tool.poetry.dependencies]
python = ">=3.10,<4.0" # for correct install verions packeges

